Refactored interfaces to live in one file
